fix: PR fixes and implement fix for middleware issue on `forRoot` and `forRootAsync`
