+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from './mikro-orm.module';
 export * from './mikro-orm.common';
 export * from './mikro-orm.middleware';
-export * from './multi-mikro-orm.module';
 export * from './multiple-mikro-orm.middleware';
 export * from './typings';
